+++ cyberark-credentialprovider-pam/SdkCredentialProviderPAM.cs
@@ -12,7 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+using Keyfactor.Logging;
 using Keyfactor.Platform.Extensions;
+using Microsoft.Extensions.Logging;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -20,27 +22,67 @@ using System.Reflection;
 
 namespace Keyfactor.Extensions.Pam.CyberArk
 {
-    public class SdkCredentialProviderPAM : IPAMProvider
+    public class SdkCredentialProviderPAM : CyberArkProvider, IPAMProvider
     {
         public string Name => "CyberArk-SdkCredentialProvider";
 
+        private readonly ILogger Logger;
+        private readonly Constants.SDK SDKConstants;
+        private readonly string ExtensionPath;
+        private readonly bool UsingFrameworkLibrary = false;
+
+        public SdkCredentialProviderPAM()
+        {
+            Logger = LogHandler.GetClassLogger<SdkCredentialProviderPAM>();
+            Logger.LogTrace($"Starting up {Name} with no constructor parameters.");
+            SDKConstants = new Constants.NetStandard();
+            Logger.LogTrace($"SDK to be targeted will be {SDKConstants.DLL}");
+
+            // when lookup path is not provided, use executing assembly location (running on UO)
+            ExtensionPath = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
+            Logger.LogInformation($"{Name} determined it will use the following directory to load the SDK: {ExtensionPath}");
+        }
+
+        public SdkCredentialProviderPAM(string extensionPath, bool useFramework = true)
+        {
+            Logger = LogHandler.GetClassLogger<SdkCredentialProviderPAM>();
+            Logger.LogTrace($"Starting up {Name} with constructor parameters provided.");
+            UsingFrameworkLibrary = useFramework;
+            if (UsingFrameworkLibrary)
+            {
+                SDKConstants = new Constants.NetFramework();
+            }
+            else
+            {
+                SDKConstants = new Constants.NetStandard();
+            }
+            Logger.LogTrace($"SDK to be targeted will be {SDKConstants.DLL}");
+
+            // Extension Path (for looking up DLL) can be passed in as a Unity Constructor parameter (running on Command)
+            ExtensionPath = extensionPath;
+            Logger.LogInformation($"{Name} determined it will use the following directory to load the SDK: {ExtensionPath}");
+        }
+
         public string GetPassword(Dictionary<string, string> instanceParameters, Dictionary<string, string> initializationInfo)
         {
-            string appId = initializationInfo["AppId"];
+            string appId = GetRequiredValue(initializationInfo, "AppId");
+            Logger.LogInformation($"Configured with Initialization Parameters: AppId = {appId}");
 
-            string safe = instanceParameters["Safe"];
-            string folder = instanceParameters["Folder"];
-            string obj = instanceParameters["Object"];
+            string safe = GetRequiredValue(instanceParameters, "Safe");
+            string folder = GetRequiredValue(instanceParameters, "Folder");
+            string obj = GetRequiredValue(instanceParameters, "Object");
+            Logger.LogInformation($"Configured with Instance Parameters: Safe = {safe} ; Folder = {folder} ; Object = {obj}");
 
-            string executingDir = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-            string dll = Path.Combine(executingDir, "NetStandardPasswordSDK.dll");
+            string dll = Path.Combine(ExtensionPath, SDKConstants.DLL);
+            Logger.LogDebug($"Loading DLL: {dll}");
             var sdk = Assembly.LoadFrom(dll);
+            Logger.LogTrace("Loaded SDK DLL.");
 
             // get the types from the dll
-            Type PasswordSDKType = sdk.GetType("CyberArk.AAM.NetStandardPasswordSDK.PasswordSDK");
-            Type PasswordRequestType = sdk.GetType("CyberArk.AAM.NetStandardPasswordSDK.PSDKPasswordRequest");
-            Type PasswordSDKExceptionType = sdk.GetType("CyberArk.AAM.NetStandardPasswordSDK.Exceptions.PSDKException");
-            Type PasswordResponseType = sdk.GetType("CyberArk.AAM.NetStandardPasswordSDK.PSDKPassword");
+            Type PasswordSDKType = sdk.GetType(SDKConstants.PasswordSDKType);
+            Type PasswordRequestType = sdk.GetType(SDKConstants.PasswordRequestType);
+            Type PasswordSDKExceptionType = sdk.GetType(SDKConstants.PasswordSDKExceptionType);
+            Type PasswordResponseType = sdk.GetType(SDKConstants.PasswordResponseType);
 
             // Create Password Request
             ConstructorInfo ctor = PasswordRequestType.GetConstructor(Type.EmptyTypes);
@@ -50,7 +92,9 @@ namespace Keyfactor.Extensions.Pam.CyberArk
             }
 
             object passwordRequest;
+            Logger.LogTrace("Attempting to invoke constructor of PSDKPasswordRequest type.");
             passwordRequest = ctor.Invoke(null);
+            Logger.LogTrace($"Constructed PSDKPasswordRequest. {passwordRequest}");
 
             PropertyInfo propertyInfo = PasswordRequestType.GetProperty("ConnectionTimeout");
             propertyInfo.SetValue(passwordRequest, 30);
@@ -73,14 +117,37 @@ namespace Keyfactor.Extensions.Pam.CyberArk
 
 
             // Sending the request to get the password
-            object passwordResponse = PasswordSDKType.GetMethod("GetPassword").Invoke(null, new object[] { passwordRequest });
+            Logger.LogTrace("Attempting to invoke GetPassword method of PasswordSDK type.");
+            object passwordResponse;
+            try
+            {
+                passwordResponse = PasswordSDKType.GetMethod("GetPassword").Invoke(null, new object[] { passwordRequest });
+            }
+            catch (TargetInvocationException ex)
+            {
+                Logger.LogError(ex.InnerException, "Error occurred when invoking GetPassword method.");
+                Logger.LogError(ex.ToString());
+                Logger.LogError(ex.InnerException.ToString());
+
+                throw ex.InnerException;
+            }
+            Logger.LogTrace("Invoked GetPassword method.");
 
 
             // Analyzing the response
             propertyInfo = PasswordResponseType.GetProperty("Content");
-            var password = (char[])propertyInfo.GetValue(passwordResponse, null);
-
-            return new string(password);
+            if (UsingFrameworkLibrary)
+            {
+                // for netframework
+                var password = (string)propertyInfo.GetValue(passwordResponse, null);
+                return password;
+            }
+            else
+            {
+                //for netstandard
+                var password = (char[])propertyInfo.GetValue(passwordResponse, null);
+                return new string(password);
+            }
         }
     }
 }